push

+++ EXE201/Repository/Interface/INotificationRepository.cs
@@ -5,8 +5,8 @@ using EXE201.Models;
 public interface INotificationRepository
 {
     Task<IEnumerable<Notification>> GetAllAsync();
-    Task<Notification> GetByIdAsync(int id);
+    Task<Notification> GetByIdAsync(long id);
     Task AddAsync(Notification notification);
     Task UpdateAsync(Notification notification);
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
 }
